Start to user/post object (and other changes)
For some code readability, i've moved to a user and post object with various properties.
This makes it slightly easier to read the code that we use, for example `x-for="g in images"` now becomes `x-for="g in post.images"` making it clear that these are the post images.
We also clean up the x-data in the top div with this, as we don't need "variables" for various properties. We now have one user and post object.

This is not completely done though. This checkin renders a normal post correctly though.